+++ src/index.ts
@@ -1,2 +1,4 @@
 export { MoneriumClient } from './client';
 export * from './types';
+export * as constants from './constants';
+export { placeOrderMessage } from './utils';